# More simplification (with a bonus performance boost)