--- a/examples/big.rs
+++ b/examples/big.rs
@@ -1,3 +1,6 @@
+// For the full copyright and license information, please view the LICENSE
+// file that was distributed with this source code.
+
 extern crate term_grid;
 use term_grid::{Alignment, Cell, Direction, Filling, Grid, GridOptions};