- Extract snippets like this into a `div_ceil` function:
```rust
let mut num_lines = self.cells.len() / num_columns;
if self.cells.len() % num_columns != 0 {
    num_lines += 1;
}
```
- Simplify the `Display` impl, by precalculating the separator and using the padding functionality of `format!`.
- Simplify `theoretical_max_num_lines` by using `enumerate` and only deal with the widths instead of the the whole `Cell`.
- Remove a clone in `width_dimensions`.
- Add an example that takes a few seconds for benchmarking.

Once I had the example, I benchmarked this against `main` and already got a >2.58x improvement!
```
Benchmark 1: target/bench/old
  Time (mean ± σ):      4.412 s ±  0.048 s    [User: 3.738 s, System: 0.670 s]
  Range (min … max):    4.354 s …  4.507 s    10 runs
 
Benchmark 2: target/bench/new
  Time (mean ± σ):      1.709 s ±  0.014 s    [User: 1.085 s, System: 0.624 s]
  Range (min … max):    1.694 s …  1.739 s    10 runs
 
Summary
  'target/bench/new' ran
    2.58 ± 0.03 times faster than 'target/bench/old'
```

EDIT: Got another performance boost from 2.1x to 2.5x faster