@tertsdiepraam 2022-11-20
Note entirely sure what's going on but there's some weird things happening in this branch when I try to run all the firefox files.