@tertsdiepraam 2023-08-12
Alright, I'm putting this up for review anyway. The reason is that I want to have these cleanups before I start working on more fixes. It doesn't have to be perfect and we can fix more along the way.